% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_escape_prob.R
\name{calc_escape_prob}
\alias{calc_escape_prob}
\title{Replicate simulation and detection}
\usage{
calc_escape_prob(
  init_dat = NULL,
  surv_locs = NULL,
  sdm = NULL,
  N_seed = 1,
  min_surv_locs = 9,
  num_replications = 2,
  rand.walk = TRUE,
  step_size_os = 100,
  step_size_ad = 100,
  Time = 10,
  K = 1000,
  age_mu = 1,
  offspr_mu = 0,
  bbox = c(-800, -800, 800, 800),
  cell_res = 10,
  sdm_og = 0,
  p_alpha = 1,
  p_beta = 1,
  allow_leave = FALSE,
  attractive_areas = FALSE,
  survive_prob = FALSE,
  crw = FALSE,
  sigma = NULL,
  theta = NULL,
  random_length = FALSE,
  PLOT.IT = FALSE,
  g0 = 1,
  lam = 1/50,
  sig = 1,
  det_func = "Manouk",
  use_manouk_error = FALSE,
  run_surveil = FALSE,
  return_sim = FALSE,
  return_all_prob = FALSE,
  ...
)
}
\arguments{
\item{init_dat}{dtest}
}
\description{

}
\details{

}
