% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_offspring.R
\name{gen_offspring}
\alias{gen_offspring}
\title{Generating Offspring}
\usage{
gen_offspring(
  dat = NULL,
  step_size_os = 5,
  offspr_mu = 2,
  K = 10,
  sigma = NULL,
  theta = NULL,
  random_length = FALSE,
  sdm = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{

}
\details{

}
