% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_dectect_one.R
\name{p_detect_one}
\alias{p_detect_one}
\title{Probability to detect one individual}
\usage{
p_detect_one(
  sim = NULL,
  surv_locs = NULL,
  g0 = 1,
  lam = 1/10,
  sig = 1,
  det_func = "Manouk",
  use_manouk_error = FALSE,
  run_surveil = FALSE
)
}
\arguments{
\item{sim}{}

\item{run_surveil}{}
}
\description{

}
\details{

}
