% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_surv.R
\name{p_halfnorm}
\alias{p_halfnorm}
\title{Half normal detection function}
\usage{
p_halfnorm(d, g0 = 0.7, sig = 1)
}
\arguments{
\item{sig}{}
}
\description{
Calculates the half-normal detection function
}
\details{
Function that returns detection based on half-normal detection function
}
