% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_surv.R
\name{p_manouk}
\alias{p_manouk}
\title{Manouk's detection function}
\usage{
p_manouk(d, g0 = 1, lam = 1/10)
}
\arguments{
\item{lam}{}
}
\description{
Calculates Manouk's detection function
}
\details{
Function that returns detection based on Manouk's detection function
}
