% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_surv.R
\name{run_surv}
\alias{run_surv}
\title{Run Surveillance}
\usage{
run_surv(
  sim = NULL,
  surv_pts = NULL,
  g0 = 0.5,
  sig = 1,
  lam = 1/10,
  det_func = "Manouk"
)
}
\arguments{
\item{sim}{a simulation object generated by \code{sim_spread()}}

\item{surv_pts}{a data frame of surveillance points}

\item{g0, sig, lam}{parameters to pass to detection functions based on
\code{det_func} argument.}

\item{det_func}{a string indicating which detection function to use.  Either
"Manouk" or "HalfNorm".  "Manouk" calls the function \code{p_manouk()}.  "HalfNorm"
calls \code{p_halfnorm()}.  See help from those functions for details.}
}
\value{
\item{\code{surv_res}}{A list showing the positive traps for each individual.}
\item{\code{surv_pts}}{A data frame containing the locations of the traps.}
\item{\code{sdm}}{A raster containing the sdm of the simulation.}
}
\description{
Helper function to run surveillance. Note this function is not currently
implemented in the rest of the package.
}
\details{
Function returns results of surveillance from a simulation and specified
surveillance points.
}
