% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_capture.R
\name{sim_capture}
\alias{sim_capture}
\title{Simulation of spread and capture of pests}
\usage{
sim_capture(
  init_dat = NULL,
  N_seed = 2,
  rand.walk = TRUE,
  surv_loc = NULL,
  ntraps = NULL,
  prop_inside = 0.05,
  det_func = "Manouk",
  g0 = 1,
  lam = 1/10,
  sig = 1,
  use_manouk_error = FALSE,
  step_size_os = 100,
  step_size_ad = 50,
  Time = 10,
  K = 1000,
  age_mu = 1,
  offspr_mu = 0,
  bbox = c(0, 0, 1608, 1608),
  cell_res = 10,
  sdm = NULL,
  sdm_og = 0,
  p_alpha = 1,
  p_beta = 1,
  allow_leave = FALSE,
  crw = FALSE,
  sigma = NULL,
  theta = NULL,
  random_length = FALSE,
  attractive_areas = FALSE,
  survive_prob = FALSE,
  ...
)
}
\arguments{
\item{init_dat}{data frame object containing columns named x, y, Fate, and
Age of the initial locations of the population to simulate a spread. If
\code{NULL} (default), then initial data frame is determined by \code{N_seed} and
\code{bbox} parameters.}

\item{N_seed}{integer of number of individuals to spread if \code{init_dat} is \code{NULL}}

\item{rand.walk}{logical; if \code{TRUE}, then individuals spread via a random walk
process.}

\item{surv_loc}{a data frame containing the locations of the traps. Default
\code{NULL}.}

\item{ntraps}{Must be specified if \code{surv_loc} is \code{NULL}, the number of traps
to be generated in a grid.}

\item{prop_inside}{If \code{surv_loc} is \code{NULL}, specifies the distance from the
border of the raster the grid starts, as a proportion of the length of
the edge of the bounding box.}

\item{det_func}{a string indicating which detection function to use.  Either
"Manouk" or "HalfNorm".  "Manouk" calls the function \code{p_manouk()}.  "HalfNorm"
calls \code{p_halfnorm()}.  See help from those functions for details.}

\item{g0, sig, lam}{parameters to pass to detection functions based on
\code{det_func} argument.}

\item{use_manouk_error}{logical. If \code{TRUE} the mean probability of
non-detected individuals is used in the detection probability.}

\item{step_size_os}{step size for dispersal distances of offspring.}

\item{step_size_ad}{step size for dispersal distances of adults.}

\item{Time}{the number of time steps for the simulation.}

\item{K}{number for the carrying capacity of cells in the raster.}

\item{age_mu}{parameter to randomly assign initial ages.}

\item{offspr_mu}{parameter passed to \code{gen_offpring()} indicating the mean
number of offspring generated per individual per time step using a Poisson
distribution.}

\item{bbox}{a vector of length 4 of the extent if \code{sdm} is not given.}

\item{cell_res}{resolution of the raster if \code{sdm} not given.}

\item{sdm}{a raster of an sdm with cells between 0 and 1.  Default is \code{NULL}.}

\item{sdm_og}{value of sdm off the grid.  If zero, then individuals die off
grid.}

\item{p_alpha}{alpha parameter for Beta distribution for survival probability
(currently not implemented).}

\item{p_beta}{beta parameter for Beta distribution for survival probability
(currently not implemented).}

\item{allow_leave}{logical indicating if individuals can leave the grid.}

\item{crw}{logical indicating a correlated random walk.}

\item{sigma}{variance of random walk if specified. Default \code{NULL}}

\item{theta}{angle of random walk if specified.  Default \code{NULL}}

\item{random_length}{logical. If \code{TRUE}, random walk allows random length up
to \code{step_size_os}. Default \code{FALSE}.}

\item{attractive_areas}{logical. Passed to \code{rand_walk()}. If \code{TRUE}, specify
areas in sdm to be more attractive and prevent individuals leaving once they
enter area as defined by raster values.}

\item{survive_prob}{logical. If \code{TRUE} allow for survival/mortality based on
sdm values.}

\item{...}{additional arguments to be passed to helper functions.}
}
\value{
\item{dat}{A list of data frames, where each element of the list
corresponds to a time step in the simulated spread.  Each data
frame contains columns for the location of individuals (x, y),
the survival (Fate), age at specified time step (Age), sdm
value (sdm), and density (dens).}
\item{sdm}{A raster containing the sdm of the simulation.}
\item{surv_loc}{A data frame containing the locations of the traps.}
\item{captured}{A vector of length Time+1 of the number of captured pests
during that time step. First element is time point 0 and
always 0.}
\item{total_captured}{A vector of length Time+1 of the cumulative total
caught pests up to that point.First element is time point 0
and always 0.}
}
\description{
This function simulates spread and captures of an outbreak.
}
\details{
Takes in an sdm, initial data, and potential survey locations.  If no sdm,
data, or survey locations are specified, generates a grid of monitoring
devices in a square uniform block.  User must specify the number of traps
if no survey locations given. It is recommended to run this function multiple
times.
}
