% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_escape_prob.R
\name{calc_escape_prob}
\alias{calc_escape_prob}
\title{Replicate simulation and detection}
\usage{
calc_escape_prob(
  init_dat = NULL,
  surv_locs = NULL,
  sdm = NULL,
  N_seed = 1,
  min_surv_locs = 9,
  num_replications = 2,
  rand.walk = TRUE,
  step_size_os = 100,
  step_size_ad = 100,
  Time = 10,
  K = 1000,
  age_mu = 1,
  offspr_mu = 0,
  bbox = c(-800, -800, 800, 800),
  cell_res = 10,
  sdm_og = 0,
  p_alpha = 1,
  p_beta = 1,
  allow_leave = FALSE,
  attractive_areas = FALSE,
  survive_prob = FALSE,
  crw = FALSE,
  sigma = NULL,
  theta = NULL,
  random_length = FALSE,
  PLOT.IT = FALSE,
  g0 = 1,
  lam = 1/50,
  sig = 1,
  det_func = "Manouk",
  use_manouk_error = FALSE,
  get_first_detect = FALSE,
  run_surveil = FALSE,
  return_sim = FALSE,
  return_all_prob = FALSE,
  ...
)
}
\arguments{
\item{init_dat}{data frame object containing columns named x, y, Fate, and
Age of the initial locations of the population to simulate a spread. If
\code{NULL} (default), then initial data frame is determined by \code{N_seed} and
\code{bbox} parameters.}

\item{surv_locs}{a data frame containing the locations of the traps.}

\item{sdm}{a raster of an sdm with cells between 0 and 1.  Default is \code{NULL}.}

\item{N_seed}{integer of number of individuals to spread if \code{init_dat} is \code{NULL}}

\item{min_surv_locs}{the minimum number of trap locations to put in the sdm.}

\item{num_replications}{the number of times to replicate the simulation
spread.}

\item{rand.walk}{logical; if \code{TRUE}, then individuals spread via a random walk
process.}

\item{step_size_os}{step size for dispersal distances of offspring.}

\item{step_size_ad}{step size for dispersal distances of adults.}

\item{Time}{the number of time steps for the simulation.}

\item{K}{number for the carrying capacity of cells in the raster.}

\item{age_mu}{parameter to randomly assign initial ages.}

\item{offspr_mu}{parameter passed to \code{gen_offpring()} indicating the mean
number of offspring generated per individual per time step using a Poisson
distribution.}

\item{bbox}{a vector of length 4 of the extent if \code{sdm} is not given.}

\item{cell_res}{resolution of the raster if \code{sdm} not given.}

\item{sdm_og}{value of sdm off the grid.  If zero, then individuals die off
grid.}

\item{p_alpha}{alpha parameter for Beta distribution for survival probability
(currently not implemented).}

\item{p_beta}{beta parameter for Beta distribution for survival probability
(currently not implemented).}

\item{allow_leave}{logical indicating if individuals can leave the grid.}

\item{attractive_areas}{logical. Passed to \code{rand_walk()}. If \code{TRUE}, specify
areas in sdm to be more attractive and prevent individuals leaving once they
enter area as defined by raster values.}

\item{survive_prob}{logical. If \code{TRUE} allow for survival/mortality based on
sdm values.}

\item{crw}{logical indicating a correlated random walk.}

\item{sigma}{variance of random walk if specified. Default \code{NULL}}

\item{theta}{angle of random walk if specified.  Default \code{NULL}}

\item{random_length}{logical. If \code{TRUE}, random walk allows random length up
to \code{step_size_os}. Default \code{FALSE}.}

\item{PLOT.IT}{logical. If \code{TRUE} plot an animation of the spread simulation.
Not recommended for large or long simulations.}

\item{g0, sig, lam}{parameters to pass to detection functions based on
\code{det_func} argument.}

\item{det_func}{a string indicating which detection function to use.  Either
"Manouk" or "HalfNorm".  "Manouk" calls the function \code{p_manouk()}.  "HalfNorm"
calls \code{p_halfnorm()}.  See help from those functions for details.}

\item{use_manouk_error}{logical. If \code{TRUE} the mean probability of
non-detected individuals is used in the detection probability.}

\item{get_first_detect}{logical. If \code{TRUE} returns a vector length
\code{num_replications} of first time-point a pest was detected.  If no pest
detected in replication, value returned is 0. Returned as \code{first_detect}.
Default \code{FALSE}.}

\item{run_surveil}{logical. If \code{TRUE}, returns surveillance of individuals.}

\item{return_sim}{logical. If \code{TRUE}, returns a list of lists where each
element of the list is a replication generated by \code{sim_spread()}.  Returned
as \code{sim}. Default \code{FALSE}.}

\item{return_all_prob}{logical. If \code{TRUE}, then a \code{num_replication} by \code{Time}
matrix is returned as \code{probs}.  Each row is one simulation spread. Default
\code{FALSE}.}

\item{...}{additional arguments to be passed to helper functions.}
}
\value{
\item{\code{mean_prob}}{A vector of length \code{Time + 1} indicating the mean estimated
probability of detecting at least one individual at each
time-step.}
\item{\code{probs}}{If \code{return_all_prob=TRUE}, a matrix of \code{num_replictaions} rows
by \code{Time + 1} columns of the detection probability of one
individual at each time-step for a given simulation.}
\item{\code{sim}}{If \code{return_sim=TRUE}, the list generated by \code{sim_spread()} is
returned. See help from the function \code{sim_spread} for details.}
\item{\code{first_detect}}{If\code{get_first_detect=TRUE}, a vector of length
\code{num_replications} of first time-point a pest was
detected for each replication. If no pest deteted,
element of vector is returned as 0.}
}
\description{
The main function of the \code{trapDetect} package.  This function is a wrapper to
call \code{sim_spread()} if necessary and \code{p_detect_one()} multiple times (recommended).
}
\details{
This function generates multiple simulations of a population of individuals
spreading through a given sdm (if supplied).
}
