% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_offspring.R
\name{gen_offspring}
\alias{gen_offspring}
\title{Generating Offspring}
\usage{
gen_offspring(
  dat = NULL,
  step_size_os = 5,
  offspr_mu = 2,
  K = 10,
  sigma = NULL,
  theta = NULL,
  random_length = FALSE,
  sdm = NULL,
  ...
)
}
\arguments{
\item{dat}{tmp_go}

\item{step_size_os}{tmp_go}

\item{offspr_mu}{tmp_go}

\item{K}{tmp_go}

\item{sigma}{tmp_go}

\item{theta}{tmp_go}

\item{random_length}{tmp_go}

\item{sdm}{tmp_go}

\item{...}{tmp_go}
}
\description{
TMP
}
\details{
TMP
}
