% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_offspring.R
\name{gen_offspring}
\alias{gen_offspring}
\title{Generating Offspring}
\usage{
gen_offspring(
  dat = NULL,
  step_size_os = 5,
  offspr_mu = 2,
  K = 10,
  sigma = NULL,
  theta = NULL,
  random_length = FALSE,
  sdm = NULL,
  ...
)
}
\arguments{
\item{dat}{tmp_go}

\item{step_size_os}{step size for dispersal distances of offspring.}

\item{offspr_mu}{parameter indicating the mean number of offspring generated
per individual per time step using a Poisson distribution.}

\item{K}{number for the carrying capacity of cells in the raster.}

\item{sigma}{variance of random walk if specified. Defalut \code{NULL}}

\item{theta}{angle of random walk if specified.  Default \code{NULL}}

\item{random_length}{logical. If \code{TRUE}, random walk allows random length up
to \code{step_size_os}. Default \code{FALSE}.}

\item{sdm}{ta raster of an sdm with cells between 0 and 1.  Default is \code{NULL}.}

\item{...}{additional arguments to be passed to other functions.}
}
\description{
TMP
}
\details{
TMP
}
