% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_dectect_one.R
\name{p_detect_one}
\alias{p_detect_one}
\title{Probability to detect one individual}
\usage{
p_detect_one(
  sim = NULL,
  surv_locs = NULL,
  g0 = 1,
  lam = 1/10,
  sig = 1,
  det_func = "Manouk",
  use_manouk_error = FALSE,
  run_surveil = FALSE
)
}
\arguments{
\item{sim}{a simulation generated by \code{sim_spread()}.}

\item{surv_locs}{a data frame containing the locations of the traps.}

\item{g0, sig, lam}{parameters to pass to detection functions based on
\code{det_func} argument.}

\item{det_func}{a string indicating which detection function to use.  Either
"Manouk" or "HalfNorm".  "Manouk" calls the function \code{p_manouk()}.  "HalfNorm"
calls \code{p_halfnorm()}.  See help from those functions for details.}

\item{use_manouk_error}{logical. If \code{TRUE} the mean probability of
non-detected individuals is used in the detection probability.}

\item{run_surveil}{logical. If \code{TRUE}, returns surveillance of individuals.}
}
\value{
If \code{run_surveil=FALSE}, only \code{p_cum_out} is returned.  Else, the following is
returned:
\item{\code{p_cum_out}}{The probability of detecting at least one individual.}
\item{\code{surv_res}}{A list showing the positive traps for each individual.}
\item{\code{surv_pts}}{A data frame containing the locations of the traps.}
\item{\code{sdm}}{A raster containing the sdm of the simulation.}
}
\description{
A function that calculates the probability of detecting at least one
individual at a given time for a given trap configuration.
}
\details{
This function requires a simulation from \code{sim_spread()} and a data frame of
locations for the traps.  It returns the probability of detecting at least
one individual from the population in the region.
}
