% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_dectect_one.R
\name{p_detect_one}
\alias{p_detect_one}
\title{Probability to detect one individual}
\usage{
p_detect_one(
  sim = NULL,
  surv_locs = NULL,
  g0 = 1,
  lam = 1/10,
  sig = 1,
  det_func = "Manouk",
  use_manouk_error = FALSE,
  run_surveil = FALSE
)
}
\arguments{
\item{sim}{a simulation generated by \code{sim_spread}.}

\item{surv_locs}{tmp_pdo}

\item{g0}{parameters to pass to detection functions based on
\code{det_func} argument.}

\item{lam}{parameters to pass to detection functions based on
\code{det_func} argument.}

\item{sig}{parameters to pass to detection functions based on
\code{det_func} argument.}

\item{det_func}{a string indicating which detection function to use.  Either
"Manouk" or "HalfNorm".  "Manouk" calls the function \code{p_manouk()}.  "HalfNorm"
calls \code{p_halfnorm()}.  See help from those functions for details.}

\item{use_manouk_error}{tmp_pdo}

\item{run_surveil}{tmp_pdo}
}
\description{
TMP
}
\details{
TMP
}
