% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_surv.R
\name{p_halfnorm}
\alias{p_halfnorm}
\title{Half Normal detection function}
\usage{
p_halfnorm(d, g0 = 0.7, sig = 1)
}
\arguments{
\item{d}{the distance of the individual to the device.}

\item{g0}{the probability of detection given the distance is 0.}

\item{sig}{the standard deviation of detection.}
}
\value{
The value of the Half Normal detection function.
}
\description{
Calculates the half-normal detection function
}
\details{
Function that returns detection based on half-normal detection function
}
