% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_surv.R
\name{p_manouk}
\alias{p_manouk}
\title{Manoukis detection functio}
\usage{
p_manouk(d, g0 = 1, lam = 1/10)
}
\arguments{
\item{d}{the distance of the individual to the device.}

\item{g0}{the probability of detection given the distance is 0.}

\item{lam}{the attractiveness of a trap, where \eqn{1/\lambda} represents a
approximately 65\% probability of detection.}
}
\description{
Calculates Manoukis detection function
}
\details{
Function that returns detection based on Manouk's detection function
}
\references{
Manoukis, N., Hall, B. & Geib, S. A Computer Model of Insect Traps in a
Landscape. Sci Rep 4, 7015 (2014). https://doi.org/10.1038/srep07015
}
