% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_walk.R
\name{rand_walk}
\alias{rand_walk}
\title{Random Walk}
\usage{
rand_walk(
  x = 0,
  y = 0,
  step_size = 1,
  sigma = NULL,
  theta = NULL,
  random_length = FALSE,
  sdm = NULL,
  attractive_areas = TRUE
)
}
\arguments{
\item{x, y}{the location (x,y) of the init}

\item{attractive_areas}{}
}
\description{

}
\details{

}
