% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_surv.R
\name{run_surv}
\alias{run_surv}
\title{Run Surveillance}
\usage{
run_surv(
  sim = NULL,
  surv_pts = NULL,
  g0 = 0.5,
  sig = 1,
  lam = 1/10,
  det_func = "Manouk"
)
}
\arguments{
\item{sim}{a simulation object generated by \code{sim_spread()}}

\item{surv_pts}{a data frame of surveillance points}

\item{g0, sig, lam}{parameters to pass to detection functions based on
\code{det_func} argument.}

\item{det_func}{a string indicating which detection function to use.  Either
"Manouk" or "HalfNorm".  "Manouk" calls the function \code{p_manouk()}.  "HalfNorm"
calls \code{p_halfnorm()}.  See help from those functions for details.}
}
\description{
Helper function to run surveillance.
}
\details{
<<<<<<< HEAD
Function returns results of surveillance from a simulation and specified
surveillance points.
=======
<<<<<<< HEAD
Function returns results of surveillance from a simulation and specified
surveillance points.
=======
Function that returns
>>>>>>> dba8220e39fbb389dee8c1ea6bb0e995bee640b1
>>>>>>> b58fced4e2dc03b08c1750c35803db4f4388f032
}
